+++ src/data/champions.ts
@@ -1,0 +1,28 @@
+export const championPokemonNames = [
+  'abomasnow','absol','aegislash','aerodactyl','aggron','alakazam','alcremie','altaria','ampharos',
+  'annihilape','appletun','araquanid','arbok','arcanine','archaludon','ariados','armarouge',
+  'aromatisse','audino','aurorus','avalugg','azumarill','banette','barbaracle','basculegion',
+  'bastiodon','beartic','beedrill','bellibolt','blastoise','blaziken','camerupt','castform',
+  'ceruledge','chandelure','charizard','chesnaught','chimecho','clawitzer','clefable','cofagrigus',
+  'conkeldurr','corviknight','crabominable','decidueye','dedenne','delphox','diggersby','ditto','dragalge',
+  'dragapult','dragonite','drampa','eelektross','emboar','emolga','empoleon','espathra','espeon',
+  'excadrill','farigiraf','feraligatr','flapple','flareon','floette','florges','forretress','froslass',
+  'furfrou','gallade','garchomp','gardevoir','garganacl','gengar','gholdengo','glaceon','glalie',
+  'glimmora','gliscor','golurk','goodra','gourgeist','greninja','grimmsnarl','gyarados','hatterene',
+  'hawlucha','heliolisk','heracross','hippowdon','houndoom','houndstone','hydrapple','hydreigon',
+  'incineroar','infernape','jolteon','kangaskhan','kingambit','kleavor','klefki','krookodile',
+  'leafeon','liepard','lopunny','lucario','luxray','lycanroc','machamp','malamar','mamoswine',
+  'manectric','maushold','mawile','medicham','meganium','meowscarada','meowstic','metagross',
+  'milotic','mimikyu','morpeko','mudsdale','musharna','ninetales','noivern','oranguru','orthworm',
+  'overqwil','palafin','pangoro','passimian','pelipper','pidgeot','pikachu','pinsir','politoed',
+  'polteageist','primarina','pyroar','quaquaval','qwilfish','raichu','rampardos','reuniclus','rhyperior',
+  'roserade','rotom','rotom-wash','rotom-heat','rotom-mow','rotom-frost','runerigus','sableye','salazzle',
+  'samurott','sandaconda','sceptile','scizor','scolipede','scovillain','scrafty','serperior','sharpedo',
+  'simipour','simisage','simisear','sinistcha','skarmory','skeledirge','slowbro','slowking','slurpuff',
+  'sneasler','snorlax','spiritomb','staraptor','starmie','steelix','stunfisk','swampert','sylveon',
+  'talonflame','tauros','tinkaton','torkoal','torterra','toucannon','toxapex','toxicroak','trevenant',
+  'tsareena','typhlosion','tyranitar','tyrantrum','umbreon','vanilluxe','vaporeon','venusaur','victreebel',
+  'vileplume','vivillon','volcarona','watchog','weavile','whimsicott','wyrdeer','zoroark','kommo-o'
+];
+
+export const championFormPatterns = ['mega-', 'rotom-', '-alola', '-galar', '-hisui'];